# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/bookings": {
      "post": {
        "operationId": "BookingsController_bookByDate",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      },
      "get": {
        "operationId": "BookingsController_getUserBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/pay/{bookingId}": {
      "post": {
        "operationId": "BookingsController_initiatePayment",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/bookings/user/{userId}": {
      "get": {
        "operationId": "BookingsController_getUserBookingsByAdmin",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/bookings/{bookingId}/cancel": {
      "patch": {
        "operationId": "BookingsController_cancelSingleBooking",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/bookings/admin/book": {
      "post": {
        "operationId": "BookingsController_adminBookForUser",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/bookings/admin/book-cash": {
      "post": {
        "operationId": "BookingsController_adminBookCash",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
...
    "/api/investments/plans": {
      "get": {
        "operationId": "InvestmentsController_getInvestmentPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Investment plans retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available investment plans",
        "tags": [
          "Investments"
        ]
      }
    },
    "/api/investments/invest": {
      "post": {
        "operationId": "InvestmentsController_invest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investment created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Investment plan not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invest in a plan",
        "tags": [
          "Investments"
        ]
      }
    },
    "/api/investments/my-investments": {
      "get": {
        "operationId": "InvestmentsController_getMyInvestments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User investments retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user investments",
        "tags": [
          "Investments"
        ]
      }
    },
    "/api/investments/active": {
      "get": {
        "operationId": "InvestmentsController_getActiveInvestments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active investments retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get active investments",
        "tags": [
          "Investments"
        ]
      }
    },
    "/api/investments/history": {
      "get": {
        "operationId": "InvestmentsController_getInvestmentHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Investment history retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment history",
        "tags": [
          "Investments"
        ]
      }
    },
    "/api/investments/summary": {
      "get": {
        "operationId": "InvestmentsController_getInvestmentSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Investment summary retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment summary",
        "tags": [
          "Investments"
        ]
      }
    },
    "/api/investments/{id}": {
      "get": {
        "operationId": "InvestmentsController_getInvestmentById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Investment retrieved successfully"
          },
          "404": {
            "description": "Investment not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment by ID",
        "tags": [
          "Investments"
        ]
      }
    },
...
    "/api/admin/investments/plans": {
      "post": {
        "description": "As an Admin, I want to create new investment plans so users can invest. Use the example in the request body to proceed.",
        "operationId": "AdminInvestmentsController_createInvestmentPlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestmentPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Investment plan created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Plan already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new investment plan",
        "tags": [
          "Admin Investment Product Management"
        ]
      },
      "get": {
        "description": "List plans with optional filters. Use query params for status, visibility, riskLevel, category, page, limit.",
        "operationId": "AdminInvestmentsController_getAllInvestmentPlans",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "enum": [
                "active",
                "inactive",
                "archived"
              ],
              "type": "string"
            }
          },
          {
            "name": "visibility",
            "required": false,
            "in": "query",
            "description": "Filter by visibility",
            "schema": {
              "enum": [
                "public",
                "private",
                "archived"
              ],
              "type": "string"
            }
          },
          {
            "name": "riskLevel",
            "required": false,
            "in": "query",
            "description": "Filter by risk level",
            "schema": {
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH"
              ],
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "example": "Fixed Income",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 0,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Investment plans retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all investment plans",
        "tags": [
          "Admin Investment Product Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "CvestAPI",
    "description": "API documentation for CvestAPI",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvestDto": {
        "type": "object",
        "properties": {}
      },
      "CreateInvestmentPlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Plan name (e.g. Conservative 10%, Balanced 15%)",
            "example": "Balanced 15%"
          },
          "rate": {
            "type": "number",
            "description": "ROI rate in percent (0-100)",
            "example": 15,
            "minimum": 0,
            "maximum": 100
          },
          "durationDays": {
            "type": "number",
            "description": "Investment duration in days",
            "example": 30,
            "minimum": 1
          },
          "minAmount": {
            "type": "number",
            "description": "Minimum investment amount (USDT)",
            "example": 500,
            "minimum": 1
          },
          "maxAmount": {
            "type": "number",
            "description": "Maximum investment amount (USDT). Omit for no cap.",
            "example": 50000,
            "minimum": 0
          },
          "visibility": {
            "type": "string",
            "description": "Visibility of the plan",
            "enum": [
              "public",
              "private"
            ],
            "example": "public"
          },
          "maxActiveUsers": {
            "type": "number",
            "description": "Max number of users who can invest (0 = unlimited)",
            "example": 0,
            "minimum": 0
          },
          "allowMultipleInvestments": {
            "type": "boolean",
            "description": "Allow user to have multiple positions in this plan",
            "example": true
          },
          "riskLevel": {
            "type": "string",
            "description": "Risk level",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "MEDIUM"
          },
          "category": {
            "type": "string",
            "description": "Category label (e.g. Fixed Income, Growth)",
            "example": "Fixed Income"
          },
          "description": {
            "type": "string",
            "description": "Public description shown to investors",
            "example": "30-day fixed return plan with 15% annualized ROI. Ideal for balanced risk appetite."
          },
          "adminNotes": {
            "type": "string",
            "description": "Internal admin notes (not shown to users)",
            "example": "Q1 2025 launch. Review limits in June."
          }
        },
        "required": [
          "name",
          "rate",
          "durationDays",
          "minAmount"
        ]
      },
...
    }
  }
}